requirements.txt
Flask>=3.0
pycryptodome>=3.20
cryptography>=42.0
Werkzeug>=3.0
